Check the SEM output in semlbci

Check whether the model and the estimation method in the sem_out object passed to semlbci() are supported by the current version of semlbci().

check_sem_out(sem_out)

Arguments

sem_out
The output from an SEM analysis. Currently only supports a lavaan::lavaan object.

Value

A numeric vector of one element. If 0, the model and estimation method are officially supported. If larger than zero, then the model and method are not officially supported. If less than zero, then the model and/or the method are officially not supported.

The attributes info contains the reason for a value other than zero.

Details

This function is to be used by semlbci() but is exported such that the compatibility of an SEM output can be checked independently.

Estimation methods (estimator in lavaan::lavaan()) currently supported:

Maximum likelihood (ML)

Full information maximum likelihood with missing data (fiml)

Generalized least squares (GLS)

Weighted least squares (a.k.a. asymptotically distribution free) (WLS)

Estimation methods not yet supported:

Unweighted least squares (ULS)

Diagonally weighted least squares (DWLS)

Variants with robust standard errors and/or robust test statistics:

MLM, MLMV, MLMVS, MLF, and MLR.

WLSM, WLSMV.

ULSM, ULSMV.

Currently supported models:

Single group models with continuous variables.

Models not tested:

Models with categorical variables.

Models not yet supported:

Multigroup models.

Models with formative factors

Multilevel models

See also

Examples

library(lavaan)

#> This is lavaan 0.6-8
#> lavaan is FREE software! Please report any bugs.

data(cfa_two_factors)
mod <-
"
f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
"

fit <- sem(mod, cfa_two_factors)

# Should be 0
check_sem_out(fit)

#> [1] 0

fit2 <- sem(mod, cfa_two_factors, estimator = "MLR")

# Should be negative because MLR is officially not supported
check_sem_out(fit2)

#> [1] -2
#> attr(,"info")
#> [1] "Standard error method robust.huber.white is not yet supported."
#> [2] "Test method yuan.bentler.mplus is not yet supported."