Find the lower or upper bound for one parameter by the Neale-Miller-1997 approach

Find the lower or upper bound of the likelihood-based confidence interval (LBCI) for one parameter by the Neale-Miller-1997 approach.

ci_bound_nm_i(
  i = NULL,
  npar = NULL,
  sem_out = NULL,
  f_constr = NULL,
  which = NULL,
  history = FALSE,
  perturbation_factor = 0.9,
  lb_var = -Inf,
  wald_ci_start = TRUE,
  standardized = FALSE,
  opts = list(),
  test_generic = TRUE,
  ciperc = 0.95,
  ci_limit_ratio_tol = 1.5,
  verbose = FALSE,
  ...
)

Arguments

i
The position of the target parameters as in the parameter table of lavaan as generated by lavaan::parameterTable().

npar
The number of free parameters, including those constrained to be equal.

sem_out
The fit object. Currently supports a lavaan::lavaan object only.

f_constr
The constraint function generated by set_constraint_nm. This argument is no longer used because the constraint function will be generated internally.

which
Whether the lower bound or the upper bound is to be found. Must be "lbound" or "ubound".

history
If TRUE, return the full optimization output as an attribute. Default is FALSE.

perturbation_factor
A number multiplied to the parameter estimates in sem_out. Using the parameter estimates as starting values may lead to errors in the first few iterations. Default is .90. This argument is ignored is wald_ci_start is `TRUE.

lb_var
The lower bound for free parameters that are variances. Default is -Inf. This is not an admissible value but seems to be necessary when we need to find the LBCI bound for a user-defined parameter.

wald_ci_start
If TRUE and there are no equality constraints in the model, Wald confidence limit will be used as the starting value for the target parameter if it is not a userd-defined paramter.

standardized
If TRUE, the LBCI is for the requested estimate in the standardized solution. Default is FALSE.

opts
Options to be passed to nloptr::nloptr(). Default is list()

test_generic
If TRUE, will use the generic functions that are used for free parameters, user defined parameters, and standardized solution. The generic functions are now the only options. Therefore, this argument will be ignored.

ciperc
The proportion of coverage for the confidence interval. Default is .95.

ci_limit_ratio_tol
The tolerance for the ratio of a to b, where a is the distance between an LBCI limit and the point estimate, and the b is the distance between the original confidence limit (by default the Wald CI in lavaan::lavaan()) and the point estimate. If the ratio is larger than this value, a warning is set in the status code. Default is 1.5.

verbose
If TRUE, the function will store more diagnostic information in the attribute diag. Default is FALSE.

...
Optional arguments. Not used.

Value

A one-element numeric vector with the requested bound.

The attribute diag contains basic information in the optimization process. If verbose is TRUE, more information is stored in this attribute.

Details

NOTE

This function is not supposed to be used directly by users. It is exported such that interested users can examine how a confidence bound is found.

Algorithm

This function currently supports lavaan::lavaan() outputs only. This function implements a version of the algorithm presented in Neale and Miller (1997) that use an equality constraint (see Pritikin, Rappaport, & Miller, 2017).

Limitation(s)

This function not yet implements the method for an estimate close to an attainable bound (see Pritikin, Rappaport, & Miller, 2017).

References

Neale, M. C., & Miller, M. B. (1997). The use of likelihood-based confidence intervals in genetic models. Behavior Genetics, 27(2), 113–120. https://doi.org/10.1023/A:1025681223921

Pritikin, J. N., Rappaport, L. M., & Neale, M. C. (2017). Likelihood-based confidence intervals for a parameter with an upper or lower bound. Structural Equation Modeling: A Multidisciplinary Journal, 24(3), 395–401. https://doi.org/10.1080/10705511.2016.1275969

See also

Examples

data(simple_med)
dat <- simple_med
mod <-
"
m ~ x
y ~ m
"
fit_med <- lavaan::sem(mod, simple_med, fixed.x = FALSE)

out1l <- ci_bound_nm_i(1, 5, sem_out = fit_med, which = "lbound")
out1l

#> [1] 0.8277017
#> attr(,"diag")
#> attr(,"diag")$status
#> [1] 0
#> 
#> attr(,"diag")$est_org
#> [1] 1.676127
#> 
#> attr(,"diag")$ci_org_limit
#> [1] 0.8317719
#> 
#> attr(,"diag")$ci_limit_ratio
#> [1] 1.004821
#> 
#> attr(,"diag")$fit_post_check
#> [1] TRUE
#> 
#> attr(,"diag")$start_values
#> [1] 0.8317719
#> 
#> attr(,"diag")$bound_unchecked
#> [1] 0.8277017
#>