Find the LBCI for one parameter

Find the likelihood-based confidence interval (LBCI) for one parameter.

ci_i(i, method = "wn", ...)

Arguments

i
The position of the target parameters as appeared in the parameter table of the lavaan::lavaan object.

method
The approach to be used. Can be "wn" (Wu-Neale-2012) or "nm" (Neale-Miller-1997). Default is "wn". (This default will be changed to "nm" in the future. The main function, semlbci(), which calls this function, already use "nm" as the default.)

...
Arguments to be passed to ci_bound_i(), ci_bound_nm_i(), or similar funtions to be developed.

Value

A numeric vector of two elements. The first element is the lower bound, and the second element is the upper bound.

The diagnostic information from the function called in finding the lower and upper founds are stored in the attributes lb_diag and ub_diag, respectively.

Details

This function calls a function to find a bound (ci_bound_nm_i() by default) twice to find the two bounds for a confidence interval. The default method is the Neale-Miller-1997 method. Please refer to [ci_bound_nm_i() for further information.

This function is not supposed to be used directly by users. It is exported such that interested users can examine how a confidence bound is found.

See also

Examples

data(simple_med)

library(lavaan)
mod <- 
"
m ~ x
y ~ m
"
fit_med <- lavaan::sem(mod, simple_med, fixed.x = FALSE)

parameterTable(fit_med)

#>   id lhs op rhs user block group free ustart exo label plabel  start    est
#> 1  1   m  ~   x    1     1     1    1     NA   0         .p1.  0.000  1.676
#> 2  2   y  ~   m    1     1     1    2     NA   0         .p2.  0.000  0.535
#> 3  3   m ~~   m    0     1     1    3     NA   0         .p3. 18.669 34.710
#> 4  4   y ~~   y    0     1     1    4     NA   0         .p4. 25.405 40.119
#> 5  5   x ~~   x    0     1     1    5     NA   0         .p5.  0.935  0.935
#>      se
#> 1 0.431
#> 2 0.073
#> 3 3.471
#> 4 4.012
#> 5 0.094

# Find the LBCI for the first parameter
out <- ci_i(1, npar = 5, sem_out = fit_med, method = "nm")
out[1:2]

#> [1] 0.8277017 2.5245536