LB Confidence Limit for One Parameter

Finds the likelihood-based confidence limit for one parameter.

ci_i_one(
  i,
  which = NULL,
  sem_out,
  method = "wn",
  standardized = FALSE,
  robust = "none",
  sf_full = NA,
  sf_args = list(),
  sem_out_name = NULL,
  ...
)

Arguments

i

The position of the target parameters as appeared in the parameter table of the lavaan::lavaan object.

which

Whether the lower bound or the upper bound is to be found. Must be "lbound" or "ubound".

sem_out

The SEM output. Currently supports lavaan::lavaan outputs only.

method

The approach to be used. Default is "wn" (Wu-Neale-2012 Method). The other methods are disabled for now.

standardized

Boolean. Whether the LBCI for the standardized solution is to be searched. Default is FALSE.

robust

Whether the LBCI based on robust likelihood ratio test is to be found. Only "satorra.2000" in lavaan is supported for now. If `"none"``, the default, then likelihood ratio test based on maximum likelihood estimation will be used.

sf_full

A list with the scaling factors. Ignored if robust is "none". If robust is "satorra.2000" and sf_full is supplied, then its value will be used. If robust is "satorra.2000" but sf_full is NA, then scaling factors will be computed internally.

sf_args

The list of arguments to be used for computing scaling factors if robust is "satorra.2000". Used only by semlbci().

sem_out_name

The name of the object supplied to sem_out. NULL by default. To be used by internal functions.

...

Arguments to be passed to the function corresponds to the requested method (ci_bound_wn_i() for "wn").

Value

A numeric vector of two elements. The first element is the lower bound, and the second element is the upper bound. The diagnostic information from the function called in finding the lower and upper founds are stored in the attributes lb_diag and ub_diag, for the lower bound and the upper bound, respectively.

Details

ci_i_one() calls a function (ci_bound_wn_i() by default) once to find the bound (limit) for a confidence interval. The default method is the Wu-Neale-2012 method. Please refer to ci_bound_wn_i() for further information.

This function is not supposed to be used directly by users. It is exported such that interested users can examine how a confidence bound it found.

See also

Examples

data(simple_med)

library(lavaan)
mod <-
"
m ~ x
y ~ m
"
fit_med <- lavaan::sem(mod, simple_med, fixed.x = FALSE)

parameterTable(fit_med)
#>   id lhs op rhs user block group free ustart exo label plabel  start    est
#> 1  1   m  ~   x    1     1     1    1     NA   0         .p1.  1.676  1.676
#> 2  2   y  ~   m    1     1     1    2     NA   0         .p2.  0.535  0.535
#> 3  3   m ~~   m    0     1     1    3     NA   0         .p3. 34.710 34.710
#> 4  4   y ~~   y    0     1     1    4     NA   0         .p4. 40.119 40.119
#> 5  5   x ~~   x    0     1     1    5     NA   0         .p5.  0.935  0.935
#>      se
#> 1 0.431
#> 2 0.073
#> 3 3.471
#> 4 4.012
#> 5 0.094

# Find the LBCI for the first parameter
# THe method "wn" need the constraint function:
fn_constr0 <- set_constraint(fit_med)
# Call ci_i to find the bounds
out <- ci_i_one(1, npar = 5, which = "lbound",
            sem_out = fit_med, method = "wn",
            f_constr = fn_constr0)
out$bounds
#>    lbound 
#> 0.8277016