Likelihood-Based Confidence Interval

Finds the likelihood-based confidence intervals (LBCIs) for selected free parameters in an SEM output.

semlbci(
  sem_out,
  pars = NULL,
  ciperc = 0.95,
  standardized = FALSE,
  method = "wn",
  robust = "none",
  ...,
  parallel = FALSE,
  ncpus = 2,
  use_pbapply = TRUE
)

Arguments

sem_out

The SEM output. Currently supports lavaan::lavaan outputs only.

pars

The positions of the parameters for which the LBCIs are to be searched. Use the position as appeared on the parameter tables of the sem_out. If NULL, the default, then LBCIs for all free parameters will be searched. Can also be a vector of strings to indicate the parameters on the parameter table. The parameters should be specified in lavaan::lavaan() syntax. The vector of strings will be converted by syntax_to_i() to parameter positions.

ciperc

The proportion of coverage for the confidence interval. Default is .95, requesting a 95 percent confidence interval.

standardized

If TRUE, the LBCI is for the standardized estimates.

method

The approach to be used to search for the confidence limits. Currently only "wn" (Wu-Neale-2012), the default, is supported.

robust

Whether the LBCI based on robust likelihood ratio test is to be found. Only "satorra.2000" in lavaan::lavTestLRT() is supported for now, implemented by the method proposed by Falk (2018). If "none", the default, then likelihood ratio test based on maximum likelihood estimation will be used.

...

Arguments to be passed to ci_bound_wn_i().

parallel

If TRUE, will use parallel to parallelize the search.

ncpus

The number of workers, if parallel is TRUE. Default is 2.

use_pbapply

If TRUE, parallel is TRUE, and pbapply is installed, pbapply::pbapply() will be used to display a progress bar when finding the intervals. Default is TRUE. Ignored if parallel is FALSE.

Value

A parameter table generated by lavaan::parameterEstimates(), with the LBCIs for selected parameters added. Diagnostic information, if requested, will be included in the attributes

Details

semlbci() finds the positions of the selected parameters in the parameter table and then calls ci_i_one() once for each of them. Please see ci_i_one() and the functions it calls to find a confidence limit (currently ci_bound_wn_i()) for the technical details.

The SEM output will first be checked by check_sem_out() to see whether the model and the estimation method are supported.

Currently supports lavaan::lavaan outputs only.

References

Falk, C. F. (2018). Are robust standard errors the best approach for interval estimation with nonnormal data in structural equation modeling? Structural Equation Modeling: A Multidisciplinary Journal, 25(2), 244-266. doi:10.1080/10705511.2017.1367254

Neale, M. C., & Miller, M. B. (1997). The use of likelihood-based confidence intervals in genetic models. Behavior Genetics, 27(2), 113-120. doi:10.1023/A:1025681223921

Pritikin, J. N., Rappaport, L. M., & Neale, M. C. (2017). Likelihood-based confidence intervals for a parameter with an upper or lower bound. Structural Equation Modeling: A Multidisciplinary Journal, 24(3), 395-401. doi:10.1080/10705511.2016.1275969

Examples

library(lavaan)
mod <-
"
m ~ a*x
y ~ b*m
ab := a * b
"
fit_med <- sem(mod, simple_med, fixed.x = FALSE)
p_table <- parameterTable(fit_med)
p_table
#>   id lhs op rhs user block group free ustart exo label plabel  start    est
#> 1  1   m  ~   x    1     1     1    1     NA   0     a   .p1.  1.676  1.676
#> 2  2   y  ~   m    1     1     1    2     NA   0     b   .p2.  0.535  0.535
#> 3  3   m ~~   m    0     1     1    3     NA   0         .p3. 34.710 34.710
#> 4  4   y ~~   y    0     1     1    4     NA   0         .p4. 40.119 40.119
#> 5  5   x ~~   x    0     1     1    5     NA   0         .p5.  0.935  0.935
#> 6  6  ab := a*b    1     0     0    0     NA   0    ab         0.000  0.897
#>      se
#> 1 0.431
#> 2 0.073
#> 3 3.471
#> 4 4.012
#> 5 0.094
#> 6 0.261
lbci_med <- semlbci(fit_med,
                    pars = c("m ~ x",
                             "y ~ m",
                             "ab :="))
lbci_med
#> 
#> Results:
#>   id lhs op rhs label    lb   est    ub wald_l wald_u cl_lb cl_ub
#> 1  1   m  ~   x     a 0.828 1.676 2.525  0.832  2.520  0.95  0.95
#> 2  2   y  ~   m     b 0.391 0.535 0.679  0.391  0.679  0.95  0.95
#> 6  6  ab := a*b    ab 0.427 0.897 1.464  0.385  1.409  0.95  0.95
#> 
#> Annotation:
#> * lb, ub: The lower and upper likehood-based limits.
#> * est: The point estimates from the original lavaan output
#> * wald_l, wald_u: The lower and upper Wald limits, extracted from the
#>     original lavaan output.
#> * cl_lb, cl_ub: The achieved level of confidence of the bound, computed
#>     from the chi-square difference test. Should be close to the
#>     requested level of confidence.
#> 
#> Call:
#> semlbci(sem_out = fit_med, pars = c("m ~ x", "y ~ m", "ab :="))