Print Method of a semlbci Object

Prints the results of a semlbci object, the output of semlbci().

# S3 method for semlbci
print(
  x,
  digits = 3,
  annotation = TRUE,
  time = FALSE,
  verbose_if_needed = TRUE,
  drop_no_lbci = TRUE,
  ...
)

Arguments

x

The output of semlbci().

digits

The number of digits after decimal point. To be passed to round(). Default is 3.

annotation

If TRUE, print table notes. Default is TRUE.

time

If TRUE, print the time spent on each limit. Default is FALSE.

verbose_if_needed

If TRUE, additional diagnostic information will be printed if necessary. If FALSE, additional diagnostic information will always be printed. Default is TRUE.

drop_no_lbci

If TRUE, parameters not included in the search will be removed. Default is TRUE.

...

Other arguments. They will be ignored.

Value

Nothing. Its side effect is used (printing the results).

Details

Prints the results of semlbci() as a table.

Examples

library(lavaan)
mod <-
"
m ~ a*x
y ~ b*m
ab := a * b
"
fit_med <- sem(mod, simple_med, fixed.x = FALSE)
p_table <- parameterTable(fit_med)
p_table
#>   id lhs op rhs user block group free ustart exo label plabel  start    est
#> 1  1   m  ~   x    1     1     1    1     NA   0     a   .p1.  1.676  1.676
#> 2  2   y  ~   m    1     1     1    2     NA   0     b   .p2.  0.535  0.535
#> 3  3   m ~~   m    0     1     1    3     NA   0         .p3. 34.710 34.710
#> 4  4   y ~~   y    0     1     1    4     NA   0         .p4. 40.119 40.119
#> 5  5   x ~~   x    0     1     1    5     NA   0         .p5.  0.935  0.935
#> 6  6  ab := a*b    1     0     0    0     NA   0    ab         0.000  0.897
#>      se
#> 1 0.431
#> 2 0.073
#> 3 3.471
#> 4 4.012
#> 5 0.094
#> 6 0.261
lbci_med <- semlbci(fit_med,
                    pars = c("ab :="))
lbci_med
#> 
#> Results:
#>   id lhs op rhs label    lb   est    ub wald_l wald_u cl_lb cl_ub
#> 6  6  ab := a*b    ab 0.427 0.897 1.464  0.385  1.409  0.95  0.95
#> 
#> Annotation:
#> * lb, ub: The lower and upper likehood-based limits.
#> * est: The point estimates from the original lavaan output
#> * wald_l, wald_u: The lower and upper Wald limits, extracted from the
#>     original lavaan output.
#> * cl_lb, cl_ub: The achieved level of confidence of the bound, computed
#>     from the chi-square difference test. Should be close to the
#>     requested level of confidence.
#> 
#> Call:
#> semlbci(sem_out = fit_med, pars = c("ab :="))

print(lbci_med, verbose_if_needed = FALSE)
#> 
#> Results:
#>   id lhs op rhs label    lb   est    ub ok_l ok_u wald_l wald_u ratio_l ratio_u
#> 6  6  ab := a*b    ab 0.427 0.897 1.464    0    0  0.385  1.409   0.919   1.108
#>   check_l check_u cl_lb cl_ub
#> 6    TRUE    TRUE  0.95  0.95
#> 
#> Annotation:
#> * lb, ub: The lower and upper likehood-based limits.
#> * est: The point estimates from the original lavaan output
#> * ok_l, ok_u: Whether the search encountered any problem. If no problem
#>     encountered, it is equal to 0. Any value other than 0 indicates
#>     something was wrong in the search.
#> * wald_l, wald_u: The lower and upper Wald limits, extracted from the
#>     original lavaan output.
#> * cl_lb, cl_ub: The achieved level of confidence of the bound, computed
#>     from the chi-square difference test. Should be close to the
#>     requested level of confidence.
#> * ratio_l, ratio_u: Ratio of a to b, a = Distance from the point
#>     estimate to the likehood-based limit, b = Distance from the point
#>     estimate to the Wald limit.
#> * check_l, check_u:Whether the final solution of a limit passed the
#>     post optimization check of lavaan by lavaan::lavInspect(fit,
#>     'post.check'), where fit is the final solution.
#> 
#> Call:
#> semlbci(sem_out = fit_med, pars = c("ab :="))