Find the LBCIs for selected free parameters in an SEM output

Find the likelihood-based confidence intervals (LBCIs) for selected free parameters in an SEM output

semlbci(
  sem_out,
  pars = NULL,
  ciperc = 0.95,
  standardized = FALSE,
  method = "nm",
  ...,
  parallel = FALSE,
  ncpu = 2
)

Arguments

sem_out
The SEM output. Currently supports lavaan::lavaan outputs only.

pars
The positions of the parameters for which the LBCI to be found. Use the position as appeared on the parameter tables of the sem_out. If NULL, the default, then LBCIs for all free parameters will be searched. Can also be a vector of strings to indicate the paramters on the parameter table. The parameters should be specified in lavaan::lavaan() syntax. The vector of strings will be converted by syntax_to_i() to parameter positions.

ciperc
The proportion of coverage for the confidence interval. Default is .95, requesting a 95 percent confidence interval.

standardized
If TRUE, the LBCI is for the standardized estimates.

method
The approach to be used to search for the confidence limits. Can be "wn" (Wu-Neale-2012) or "nm" (Neale-Miller-1997). Default is "nm".

...
Arguments to be passed to ci_bound_i().

parallel
If TRUE, will use parallel. Currently disabled and so this argument will be ignored.

ncpu
The number of workers, if parallel is TRUE. Currently ignored.

Value

A parameter table generated by lavaan::parameterEstimates(), with the LBCIs for selected parameters added. Diagnostic information, if requested, will be included in the attributes

Details

This function finds the positions of the selected parameters in the parameter table and then call ci_i() once for each of them. Please see ci_i() and the functions it call to find a confidence limit (currently ci_bound_nm_i() and ci_bound_i()) for the technical details.

The SEM output will first checked by check_sem_out() to see whether the model and the estimation method are supported.

Currently supports lavaan::lavaan outputs only.

References

Neale, M. C., & Miller, M. B. (1997). The use of likelihood-based confidence intervals in genetic models. Behavior Genetics, 27(2), 113–120. https://doi.org/10.1023/A:1025681223921

Pritikin, J. N., Rappaport, L. M., & Neale, M. C. (2017). Likelihood-based confidence intervals for a parameter with an upper or lower bound. Structural Equation Modeling: A Multidisciplinary Journal, 24(3), 395–401. https://doi.org/10.1080/10705511.2016.1275969

Examples

if (FALSE) {
library(lavaan)
mod <-
"
m ~ a*x
y ~ b*m
ab := a * b
"
fit_med <- sem(mod, simple_med, fixed.x = FALSE)
p_table <- parameterTable(fit_med)
p_table
lbci_med <- semlbci(fit_med, 
                    pars = c("m ~ x",
                             "y ~ m",
                             "ab :="),
                    method = "nm")
lbci_med
}