Constraint for Finding the LBCI by the Wu-Neale-2012 Approach

Sets the constraint for finding the likelihood-based confidence interval by the Wu-Neale-2012 approach.

set_constraint(sem_out, ciperc = 0.95)

Arguments

sem_out

The SEM output. Currently supports lavaan::lavaan outputs only.

ciperc

The probability of coverage for the confidence interval. Default is .95.

Value

A constraint function for nloptr.

Details

The Wu-Neale-2012 approach uses a simple objective function that is optimized with a constraint. set_constraint generates the constraint function used by ci_bound_wn_i().

Currently supports lavaan::lavaan outputs only.

This function is not to be used by normal users.

Examples

library(lavaan)
data(simple_med)
dat <- simple_med
mod <-
"
m ~ x
y ~ m
"
fit_med <- sem(mod, simple_med, fixed.x = FALSE)

fn_constr0 <- set_constraint(fit_med)
out <- fn_constr0(coef(fit_med), sem_out = fit_med)
out
#> $objective
#> [1] 0.02637152
#> 
#> $gradient
#>              [,1]         [,2]          [,3]          [,4] [,5]
#> [1,] 6.938894e-18 6.938894e-18 -5.421011e-20 -2.207298e-18    0
#> 
#> $constraints
#> [1] -0.009603647
#> 
#> $jacobian
#>              [,1]         [,2]          [,3]          [,4] [,5]
#> [1,] 6.938894e-18 6.938894e-18 -5.421011e-20 -2.207298e-18    0
#> 
#> $parameterTable
#>   id lhs op rhs block user free ustart exo label group plabel  start    est se
#> 1  1   m  ~   x     1    1    1     NA   0           1   .p1.  1.676  1.676 NA
#> 2  2   y  ~   m     1    1    2     NA   0           1   .p2.  0.535  0.535 NA
#> 3  3   m ~~   m     1    0    3     NA   0           1   .p3. 34.710 34.710 NA
#> 4  4   y ~~   y     1    0    4     NA   0           1   .p4. 40.119 40.119 NA
#> 5  5   x ~~   x     1    0    5     NA   0           1   .p5.  0.935  0.935 NA
#> 
lavTech(fit_med, "optim")$fx
#> [1] 0.02637152